Cognitus/core/requirements_web.txt
Django==4.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
pandas==2.0.3
psycopg2-binary==2.9.6
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
redis==4.6.0
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==2.0.19
sqlparse==0.4.4
